Extract commitToPullRequest from reviewTime
